Revert "Migration to typescript + ESLint and Prettier"
