+++ .eslintrc.yml
@@ -1,0 +1,35 @@
+env:
+  browser: true
+  es2021: true
+  jest: true
+extends:
+  - eslint:recommended
+  - plugin:react/recommended
+  - plugin:react-hooks/recommended
+  - plugin:prettier/recommended
+parser: '@typescript-eslint/parser'
+parserOptions:
+  ecmaFeatures:
+    jsx: true
+  ecmaVersion: latest
+  sourceType: module
+plugins:
+  - '@typescript-eslint'
+  - prettier
+  - react
+  - react-hooks
+rules:
+  # note you must disable the base rule as it can report incorrect errors
+  no-undef: off # Recommended by ESLint to turn off for TS projects
+  no-unused-vars: off
+  '@typescript-eslint/no-unused-vars': error
+  '@typescript-eslint/no-explicit-any': error
+  'react/display-name': off
+  'react/prop-types': error
+  'react/jsx-key': error
+  prefer-const: error
+  prettier/prettier: error
+  # suppress errors for missing 'import React' in files (no longer required since React 17)
+  react/react-in-jsx-scope: off
+settings:
+  react: { version: 'detect' }
