+++ src/components/PastEvents.jsx
@@ -1,0 +1,10 @@
+function PastEvents() {
+    return (
+        <div>
+
+
+
+            
+        </div>
+    )
+}
