--- src/components/PastEvents.tsx
@@ -1,10 +1,0 @@
-function PastEvents() {
-    return (
-        <div>
-
-
-
-            
-        </div>
-    )
-}
