+++ src/hackathonWebsites/MelbourneHack/2023/Header.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import './header.css';
+import logo from '/img/HackMelbourne2023_new_logo.png';
+
+const Header = () => {
+  return (
+    <>
+      <nav id="melbournehack2023-nav">
+        <div className="logo">
+          <img src={logo} alt="Logo" id="melbournehack2023-new-logo" />
+        </div>
+        {/* <ul>
+          <li>
+            <a href="#">About Us</a>
+          </li>
+          <li>
+            <a href="#">Schedule</a>
+          </li>
+        </ul> */}
+      </nav>
+    </>
+  );
+};
+
+export default Header;
