+++ src/hackathonWebsites/MelbourneHack/2023/globe.css
@@ -1,0 +1,12 @@
+html,
+body,
+#root {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+#root {
+    background-color: #01040C;
+}
