+++ src/routes/Sponsor.jsx
@@ -1,0 +1,16 @@
+import '../styles/sponsor.css';
+import React from 'react';
+import Sponsors from '../components/Sponsor/Sponsors';
+import SponsorUs from '../components/Sponsor/SponsorUs';
+
+function Sponsor() {
+  return (
+    <div className="sponsorsContainer App">
+      <Sponsors />
+      <hr></hr>
+      <SponsorUs />
+    </div>
+  );
+}
+
+export default Sponsor;
