+++ src/routes/root.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import AboutUs from '../components/Home/AboutUs';
+import WhatWeDo from '../components/Home/WhatWeDo';
+import MeetTheTeam from '../components/Home/MeetTheTeam';
+
+function App() {
+  return (
+    <div className="App">
+      <AboutUs />
+      <WhatWeDo />
+      <MeetTheTeam />
+    </div>
+  );
+}
+
+export default App;
